Finish!

+++ deleteaccount.php
@@ -3,43 +3,56 @@
 include("config.php");
 session_start();
 
+header("X-Frame-Options: DENY");  // Clickjacking 방지
+
 //get post parameters
 
-$user=$_POST['username'];
-$old=$_POST['oldpasswd'];
+$user = mysqli_real_escape_string($db, $_POST['login_user']);
+$old = mysqli_real_escape_string($db, $_POST['oldpasswd']);
+$csrf = mysqli_real_escape_string($db, $_POST['csrf_token']);
 
 
 //check session else redirect to login 
 
-$check=$_SESSION['login_user'];
+$check = mysqli_real_escape_string($db, $_SESSION['login_user']);
+
 if($check==NULL)
 {
-	header("Location: /vulnerable/index.html");
+	header("Location: /index.php");
 }
 
 //check values else redirect to settings page
 if($check!=NULL && ($user==NULL || $old==NULL) )
 {
-header("Location: /vulnerable/settings.php");	
+header("Location: /settings.php");	
 }
 
+// CSRF detection
+if($_SESSION['csrf'] == $csrf) {
+	// 진짜 사용자 확인
+	if($check == $user){
 
-$sql="DELETE from register where username='$user' AND password='$old'";
+		$sql="DELETE from register where username='$user' AND password='$old'";
 
-echo $sql;
-echo "</br>";
+		echo htmlentities($sql);
+		echo "</br>";
 
-$result=mysqli_query($db, $sql) or die('Error querying database.');
+		$result=mysqli_query($db, $sql) or die('Error querying database.');
 
-if( mysqli_affected_rows($db)>0)
-{
-echo "<h2>Account Deleted successfully</h2>";
-session_destroy();
+		if( mysqli_affected_rows($db) > 0)
+		{
+		echo "<h2>Account Deleted successfully</h2>";
+		session_destroy();
+		}
+		else {
+			echo "<h2>Incorrect Password</h2>";
+		}
+	} else {
+		echo "<h2>You are not authorized!</h2>";
+	}
+} else {
+	echo "<h2>CSRF detected... Get out of here!</h2>";
 }
-else {
-	echo "<h2>Incorrect Password</h2>";
-}
-
 
 mysqli_close($db);
 
@@ -50,14 +63,9 @@ mysqli_close($db);
 <body>
 </br>
 
-<script>
-if(top != window) {
-  top.location = window.location
-}
 
-</script>
-<a href="/vulnerable/settings.php" > <h3> Go back </h3> </a>
+<a href="/settings.php" > <h3> Go back </h3> </a>
 </br>
-<a href="/vulnerable/index.html" > <h3>Login page </h3> </a>
+<a href="/index.php" > <h3>Login page </h3> </a>
 </body>
 </html>
